+++ packages/ui/src/react/Form.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Form = styled.form`
+  display: flex;
+  flex-direction: row-reverse;
+  align-items: center;
+`
