+++ packages/ui/src/react/Root.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const Root = styled.div`
+  position: relative;
+  display: grid;
+  grid-gap: 1em;
+`
