+++ packages/ui/src/react/SearchIcon.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+import { Search } from 'styled-icons/fa-solid/Search'
+
+export const SearchIcon = styled(Search)`
+  width: 1em;
+  pointer-events: none;
+`
